+++ src/scorpio_pipe/ui/stage_frames_window.py
@@ -7,7 +7,8 @@ from PySide6 import QtCore, QtGui, QtWidgets
 
 from scorpio_pipe.app_paths import ensure_dir
 from scorpio_pipe.inspect import KIND_ORDER
-from scorpio_pipe.wavesol_paths import resolve_work_dir, wavesol_dir
+from scorpio_pipe.paths import resolve_work_dir
+from scorpio_pipe.wavesol_paths import wavesol_dir
 from scorpio_pipe.ui.fits_preview import FitsPreviewWidget
 from scorpio_pipe.ui.frame_browser import FrameBrowser